Draw points

+++ ZodiaClash/ZodiaClash/model.h
@@ -15,24 +15,26 @@ public:
 	void Draw();
 	void AttachTexture(Texture&);
 	void AttachTexture(char const*);
-	void SetPos(double x, double y);
-	void AddPos(double x, double y);
-	void SetRot(double rot); //INPUT IN DEGREES
-	void AddRot(double rot); //INPUT IN DEGREES
-	void SetScale(double x, double y);
+	void SetPos(float x, float y);
+	void AddPos(float x, float y);
+	void SetRot(float rot); //INPUT IN DEGREES
+	void AddRot(float rot); //INPUT IN DEGREES
+	void SetScale(float x, float y);
 private:
 	glm::vec2 pos{}; //Position in screen coordinates
 	glm::vec2 scale{}; //Scale
-	double width{};
-	double height{};
-	double rotation{}; //rotation in degrees
+	float width{};
+	float height{};
+	float rotation{}; //rotation in degrees
 
 	Texture* tex{};
 	glm::mat3 matrix{};
 	glm::vec4 color{};
 
 	//PRECOMPUTED VALUES
-	double rotationRadians{};
+	float rotationRadians{};
 };
 
+//THIS CURRENTLY CAUSES MEMORY LEAKS AS VECTOR MEMORY ISNT CLEARED
+//ALSO ITS TEMPORARY, COMPONENT TO BE ADDED TO ENTITY LIST
 extern std::vector<Model> modelList;
